+++ packages/common/utils/rollout-status-watch.js
@@ -1,6 +1,6 @@
 const { setTimeout: sleep } = require("timers/promises")
 
-const rolloutStatus = require("./rollout-status")
+const defaultRolloutStatus = require("./rollout-status")
 const defaultLogger = require("./logger")
 
 module.exports = async ({
@@ -8,13 +8,14 @@ module.exports = async ({
   selector,
   kubeconfig,
   kubecontext,
-  logger = defaultLogger,
   interceptor = {},
   rolloutStatusProcesses = {},
   checkForNewResourceInterval = 3000,
   watchingTimeout = 3600000, // 60 minutes
   surviveOnBrokenCluster = false,
   kindFilter,
+  logger = defaultLogger,
+  rolloutStatus = defaultRolloutStatus,
 }) => {
   let timeoutReached = false
   const globalTimeout = setTimeout(() => {