+++ packages/webhook/src/k8s/utils/kubectl.js
@@ -1,0 +1,11 @@
+const { ctx } = require("@modjo-plugins/core")
+
+const kubectlRetry = require("~common/utils/kubectl-retry")
+
+module.exports = async (kubectlArgs, options = {}) => {
+  const sentry = ctx.get("sentry")
+  return kubectlRetry(kubectlArgs, {
+    ...options,
+    sentry,
+  })
+}