+++ plugins/contrib/deploy-with/kapp.js
@@ -59,7 +59,6 @@ module.exports = async (
   for (const signal of signals) {
     process.on(signal, () => {
       proc.kill(signal)
-      process.exit(0)
     })
   }