+++ plugins/contrib/deploy-with/kubectl.js
@@ -8,18 +8,9 @@ const isNotDefined = (val) => val === undefined || val === null || val === ""
 const defaultTo = (val, defaultVal) => (isNotDefined(val) ? defaultVal : val)
 
 module.exports = async (deploys, options, context) => {
-  const { config, utils, manifests, dryRun, needBin } = context
+  const { config, utils, manifests, dryRun, kubectl } = context
 
-  const {
-    yaml,
-    kubectlRetry,
-    logger,
-    kubectlDeleteManifest,
-    kindIsRunnable,
-    needKubectl,
-  } = utils
-
-  await needBin(needKubectl)
+  const { yaml, logger, kubectlDeleteManifest, kindIsRunnable } = utils
 
   const { kubeconfigContext, kubeconfig } = config
 
@@ -63,6 +54,7 @@ module.exports = async (deploys, options, context) => {
     kubeconfigContext,
     retryOptions: kubectlRetryOptions,
     surviveOnBrokenCluster,
+    kubectl,
   }
 
   const forceAnnotationKey = "kontinuous/kubectl-force"
@@ -96,7 +88,7 @@ module.exports = async (deploys, options, context) => {
       --wait
       --timeout=${applyTimeout}
   `
-    return kubectlRetry(kubectlDeployCommand, {
+    return kubectl(kubectlDeployCommand, {
       kubeconfig,
       kubeconfigContext,
       stdin: yamlManifest,
@@ -105,6 +97,7 @@ module.exports = async (deploys, options, context) => {
       logError: false,
       retryOptions: kubectlRetryOptions,
       surviveOnBrokenCluster,
+      kubectl,
     })
   }
 
@@ -167,7 +160,6 @@ module.exports = async (deploys, options, context) => {
       if (stopRolloutStatus) {
         stopRolloutStatus()
       }
-      process.exit(0)
     })
   }