+++ plugins/contrib/pre-deploy/01-rancher-namespaces.js
@@ -1,7 +1,7 @@
 module.exports = async (
   manifests,
   _options,
-  { utils, config, logger, needBin }
+  { utils, config, logger, kubectl }
 ) => {
   const { kubeEnsureNamespace } = utils
   const { kubeconfig, kubeconfigContext } = config
@@ -22,11 +22,9 @@ module.exports = async (
 
   logger.debug({ namespaces }, "ensure rancher namespaces are availables")
 
-  await needBin(utils.needKubectl)
-
   await Promise.all(
     rancherNamespacesManifests.map((manifest) =>
-      kubeEnsureNamespace({ kubeconfig, kubeconfigContext, manifest })
+      kubeEnsureNamespace({ kubeconfig, kubeconfigContext, manifest, kubectl })
     )
   )
   logger.debug({ namespaces }, "namespaces ready")